+++ server/.env.txt
@@ -2,3 +2,4 @@ NODE_ENV = development
 PORT = 3001
 MONGO_URI = mongodb.net 
 JWT_SECRET = abc123
+DANGEROUSLY_DISABLE_HOST_CHECK=true
