+++ server/src/controllers/workoutScheduleController.js
@@ -1,11 +1,5 @@
-const { schedule } = require('../mock/WorkoutScheduleMockData');
-
-const schedules = [
-  {
-    userID: '1',
-    schedule,
-  },
-];
+const asyncHandler = require('express-async-handler');
+const openAI = require('../utils/openaiUtil');
 
 /**
  * @desc    get workout schedule for user (userID)
@@ -23,14 +17,28 @@ const getWorkoutScheduleByUserID = (req, res) => {
  * @route   POST /workoutSchedule
  * @access  Private
  */
-const createWorkoutSchedule = (req, res) => {
-  if (!req.body.schedule) {
-    return res.status(400).send({ message: 'Missing Payload' });
+const createWorkoutSchedule = asyncHandler(async (req, res) => {
+  const tempUser = {
+    age: 25,
+    sex: 'male',
+    weight: '186',
+    BMI: '23',
+    fitness: 'high',
+    healthConditions: 'asthma',
+    height: '180',
+    timePreference: '5 days per week',
+    durationPreference: '60',
+    equipmentAcess: 'dumbbells',
+    goal: 'weight loss',
+  };
+
+  try {
+    const schedule = await openAI.generateWorkoutSchedule(tempUser);
+    res.send(schedule);
+  } catch (e) {
+    res.sendStatus(500);
   }
-  const item = { userID: req.body.userID, schedule: req.body.schedule };
-  schedules.push(item);
-  return res.send(item);
-};
+});
 
 /**
  * @desc    update workout schedule for user (userID)
@@ -38,14 +46,16 @@ const createWorkoutSchedule = (req, res) => {
  * @access  Private
  */
 const updateUserWorkoutScheduleByUserID = (req, res) => {
-  const foundItemIndex = schedules.findIndex((item) => item.userID === req.params.userID);
+  // const foundItemIndex = schedules.findIndex((item) => item.userID === req.params.userID);
+  const foundItemIndex = null;
 
   if (foundItemIndex < 0) return res.status(404).send({ message: 'Item not found' });
   if (!req.body.schedule) {
     return res.status(400).send({ message: 'Missing payload' });
   }
-  schedules[foundItemIndex].schedule = req.body.schedule;
-  return res.send(schedules[foundItemIndex]);
+  return res.status(400).send({ message: 'Missing payload' });
+  // schedules[foundItemIndex].schedule = req.body.schedule;
+  // return res.send(schedules[foundItemIndex]);
 };
 
 module.exports = {
