+++ server/src/routes/MealSchedule.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { getMealScheduleByUser, createMealScheduleForUser, updateMealScheduleForUser } = require('../controllers/mealScheduleController');
 
 const router = express.Router();
-const { schedule } = require('./mock/MealScheduleMockData');
+// const { schedule } = require('./mock/MealScheduleMockData');
 const openAI = require('../utils/openaiUtil');
 
 /**
@@ -61,8 +61,53 @@ router.post('/:userID', async (req, res) => {
     timePreference: '5 days per week',
     durationPreference: '60',
     equipmentAcess: 'dumbbells',
-    goal: 'gaining muscle'
-  }
+    goal: 'gaining muscle',
+  };
+
+  const mealSchedule = await MealSchedule.create({
+    _userInfoID: id,
+    schedule: generatedSchedule,
+    inputs: [],
+    // schedule: mockMealSchedule,
+    // schedule: {},
+    // schedule: {
+    //   Monday: {
+    //     breakfast: 'Oatmeal with berries and a sprinkle of nuts',
+    //     snack1: 'Greek yogurt with cucumber slices',
+    //     lunch: 'Grilled chicken breast with mixed vegetables',
+    //     snack2: 'Apple slices with almond butter',
+    //     dinner: 'Salmon with quinoa and steamed broccoli',
+    //   },
+    //   Tuesday: {
+    //     breakfast: 'Egg white omelette with spinach and tomatoes',
+    //     snack1: 'Protein shake with almond milk',
+    //     lunch: 'Turkey wrap with whole wheat tortilla, lettuce, and tomato',
+    //     snack2: 'Carrot sticks with hummus',
+    //     dinner: 'Grilled lean steak with sweet potato and roasted asparagus',
+    //   },
+    //   Wednesday: {
+    //     breakfast: 'Whole grain toast with avocado and poached eggs',
+    //     snack1: 'Mixed berries',
+    //     lunch: 'Quinoa salad with grilled chicken, bell peppers, and feta cheese',
+    //     snack2: 'Celery sticks with peanut butter',
+    //     dinner: 'Baked cod with brown rice and sautéed zucchini',
+    //   },
+    //   Thursday: {
+    //     breakfast: 'Greek yogurt with granola and sliced bananas',
+    //     snack1: 'Protein bar',
+    //     lunch: 'Shrimp stir-fry with mixed vegetables and brown rice',
+    //     snack2: 'Edamame',
+    //     dinner: 'Grilled tofu with quinoa and roasted Brussels sprouts',
+    //   },
+    //   Friday: {
+    //     breakfast: 'Smoothie with spinach, banana, almond milk, and protein powder',
+    //     snack1: 'Cucumber slices with tzatziki sauce',
+    //     lunch: 'Grilled chicken breast with quinoa and steamed broccoli',
+    //     snack2: 'Mixed nuts',
+    //     dinner: 'Grilled salmon with sweet potato fries and grilled asparagus',
+    //   },
+    // },
+  });
 
   try {
     const schedule = await openAI.generateMealSchedule(tempUser);
